added basic SQL TO BE IMPROVED BY ME after dinner
